fix: 코드 리뷰 피드백 5건 반영
- [Major] cancel()에서 쿠폰·재고 복원 누락 수정 (cancelOrderInternal 추출)
- [Major] RATE 할인 시 (int) 캐스팅 소수점 절삭 수정 (Money.percentage BigDecimal 오버로드)
- [Minor] 만료/미도래 쿠폰 validateUsable 단위 테스트 추가
- [Minor] @Modifying에 clearAutomatically, flushAutomatically 옵션 추가
- [Minor] 삭제된 쿠폰 이름 빈 문자열 → "(삭제된 쿠폰)" 변경

+++ apps/commerce-api/src/main/java/com/loopers/domain/coupon/Coupon.java
@@ -106,7 +106,7 @@ public class Coupon extends BaseEntity {
         if (this.discountType == DiscountType.FIXED) {
             discount = this.discountValue;
         } else {
-            discount = orderAmount.percentage((int) this.discountValue.getAmount().longValue());
+            discount = orderAmount.percentage(this.discountValue.getAmount());
             if (this.maxDiscountAmount != null) {
                 discount = discount.min(this.maxDiscountAmount);
             }
